worker_mcp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
ollama==0.3.2
requests==2.32.3
chromadb==0.5.18
kafka-python==2.0.2
